pyproject.toml
[project]
name = "anima-engine"
version = "0.1.0"
description = "Autonomous Iteration Engine — give your project a soul"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "textual>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
anima = "anima.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anima"]

[dependency-groups]
dev = [
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["anima"]
omit = [
    "src/anima/tui/*",
    "src/anima/cli.py",
    "src/anima/__main__.py",
]

[tool.ruff]
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
